reorganized most client side scripts. About to delete drawMap.js
