+++ client/script/CONSTANTS.js
@@ -1,7 +1,9 @@
-export const FISH_VALUES = {
+const TILE_SIZE = 32
+const SEA_LEVEL = 47
+const FISH_VALUES = {
   shallows: 5,
   openOcean: 15,
   deep: 25
 }
 
-export {TILE_SIZE, SEA_LEVEL} from './drawMap'
+export {TILE_SIZE, SEA_LEVEL, FISH_VALUES}
